refactor go log
